refactor(pages):changed folder structure
