+++ src/Pages/Main/AboutUs/AboutUs.tsx
@@ -1,10 +1,10 @@
-import MainLayout from "../../Layouts/MainLayout/MainLayout";
+import MainLayout from "../../../Layouts/MainLayout/MainLayout";
 
 function AboutUs() {
   return (
     <>
       <MainLayout>
-        <div>about us page</div>;
+        <div>about us page</div>
       </MainLayout>
     </>
   );
