+++ src/Pages/Main/Home/Home.tsx
@@ -1,10 +1,10 @@
-import MainLayout from "../../Layouts/MainLayout/MainLayout";
+import MainLayout from "../../../Layouts/MainLayout/MainLayout";
 
 function Home() {
   return (
     <>
       <MainLayout>
-        <div>home page</div>;
+        <div>home page</div>
       </MainLayout>
     </>
   );
