+++ src/Pages/Main/Shop/Shop.tsx
@@ -1,10 +1,10 @@
-import MainLayout from "../../Layouts/MainLayout/MainLayout";
+import MainLayout from "../../../Layouts/MainLayout/MainLayout";
 
 function Shop() {
   return (
     <>
       <MainLayout>
-        <div>shop page</div>;
+        <div>shop page</div>
       </MainLayout>
     </>
   );
