+++ src/Pages/Main/index.tsx
@@ -1,0 +1,5 @@
+export { default as Home } from "./Home/Home";
+export { default as Shop } from "./Shop/Shop";
+export { default as Blog } from "./Blog/Blog";
+export { default as ContactUs } from "./ContactUs/ContactUs";
+export { default as AboutUs } from "./AboutUs/AboutUs";
