+++ src/Services/Utils/Data/Data.ts
@@ -2,21 +2,26 @@ export const navbarItems=[
     {
         id:1,
         name:'home',
+        route:'/',
     },
     {
         id:2,
         name:'shop',
+        route:'/shop',
     },
     {
         id:3,
         name:'blog',
+        route:'/blog',
     },
     {
         id:4,
         name:'contact us',
+        route:'/contact-us',
     },
     {
         id:5,
         name:'about us',
+        route:'/about-us',
     }
 ]
